refactor: manager에서 controller로 명칭 변경
