+++ README.md
@@ -65,9 +65,9 @@ sequenceDiagram
    3. `LottoShop` : 로또를 판매하는 로또 가게입니다.
    4. `Lotto` : 로또입니다.
    5. `LottoResult` : 로또 당첨 결과지입니다.
-3. [헬퍼 클래스](./src/helpers/)
-   1. `InputManager` : 사용자로부터 입력을 받습니다.
-   2. `OutputManager` : 사용자에게 결과를 출력합니다.
+3. [컨트롤러 클래스](./src/controllers/)
+   1. `InputController` : 사용자로부터 입력을 받습니다.
+   2. `OutputController` : 사용자에게 결과를 출력합니다.
 
 현실 세계에서 로또 가게, 로또 회사, 로또는 움직일 수 없는 무생물입니다. 하지만 소프트웨어 세계에 이러한 것들을 옮겨오면서 생명을 불어넣어주었습니다. 다시말해 소프트웨어 세계에서 로또는 로또 당첨 번호를 비교하기도 하고, 로또 결과지는 직접 당첨 결과를 합할 수 있습니다.
