+++ src/controllers/index.js
@@ -1,0 +1,2 @@
+export { default as InputController } from './InputController.js';
+export { default as OutputController } from './OutputController.js';
