--- src/helpers/index.js
@@ -1,2 +1,0 @@
-export { default as InputManager } from './InputManager.js';
-export { default as OutputManager } from './OutputManager.js';
